requirements.txt
google-cloud-bigquery>=3.11.4
psycopg[binary]>=3.1.0
sqlalchemy>=2.0.19
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
pandas>=2.0.3
numpy>=1.24.3
streamlit>=1.25.0
plotly>=5.15.0
click>=8.1.6
pyyaml>=6.0.1
python-dotenv>=1.0.0
duckdb>=0.8.1
joblib>=1.3.0